Rest-java k6 tests in CI (#12243)
Add rest-java k6 tests execution as part of the workflow in order to verify the end-to-end functionality. This will be helpful to catch some issues earlier in the testing cycle. The goal is not a complete coverage of all rest-java k6 tests, so the failing ones are excluded for now.

---------

+++ test/src/test/java/org/hiero/mirror/test/e2e/acceptance/client/TokenClient.java
@@ -75,8 +75,10 @@ public class TokenClient extends AbstractNetworkClient {
     public void clean() {
         var admin = sdkClient.getExpandedOperatorAccountId();
         log.info("Deleting {} tokens and dissociating {} token relationships", tokenIds.size(), associations.size());
-        deleteOrLogEntities(tokenIds, tokenId -> delete(admin, tokenId));
-        deleteAll(associations.keySet(), association -> dissociate(association.accountId, association.tokenId));
+        var entitiesShouldBeDeleted = deleteOrLogEntities(tokenIds, tokenId -> delete(admin, tokenId));
+        if (entitiesShouldBeDeleted) {
+            deleteAll(associations.keySet(), association -> dissociate(association.accountId, association.tokenId));
+        }
     }
 
     @Override
